- ajout du squelette de la fonctionnalité des event - ajout des event perso (feuille todo.txt pour avoir une base mais ne pas se fier à elle, il y a eu des modifications pendant le code)

+++ src/engine/data/person/vitality/Health.java
@@ -10,7 +10,7 @@ public class Health extends Vitality {
     private Boolean estMalade;
     private int derniereConsultation;
 
-    public Health(int niveau, int derniereConsultation, Boolean estMalade) {
+    public Health(double niveau, int derniereConsultation, Boolean estMalade) {
         super(niveau);
         this.derniereConsultation = derniereConsultation;
         this.estMalade = estMalade;
